Added component for simple containers
* SimpleContainerComponent is it s name. Simple containers are its game.  This is the test code from default_hands_main.tsx packaged up in a way that can theoretically be reused. No new functionality here.
* Added a render function to the component interface since containers need to render the transforms for their children.
